anvik_ai/requirements.txt
aiohttp
python-dotenv
langchain
langchain-community
langchain-google-genai
pydantic
python-docx
fastapi
uvicorn
gradio